requirements.txt
streamlit
SpeechRecognition 
gtts 
audio_recorder_streamlit
transformers
torch
google.generativeai
python-dotenv
